Move ReplayData to common/types; delete slippilabTypes
